Don't require redirect-url to be set
If redirect-url is set, the only way to support both V1 and V2
content at the same time is to set a symlink in the filer, as
the repository name is directly appended to the redirect-url.

With no redirect-url set, Crane will append /v1/{repository} or
/v2/{repository} depending on the API version used to access it.

By allowing redirect-url to be omitted, we can use default-config
Pulp installations without issue for V1 and V2 content.