listRepos: calculate image inheritance
A new field in the returned object, 'images_children', is a dict of
parent_id:children, where children is a dict in the same form. This
represents the ancestry tree for docker_image units in the repository.
When an image has no children, the value is an empty dict.

The top-level keys in this dict have no parents and are the top-level
base images.

This is needed in order to implement garbage collection (OSBS-2596).

Signed-off-by: Tim Waugh <twaugh@redhat.com>