Report error when --server isn't specified (#212)
'qa' was the default value of server. It could be confusing if it's
not noticed. E.g. a user may login to the default environment without
specifying --server, but not aware it's 'qa', then get error of 'need
to log in' when doing other operations with a --server other than 'qa'.
Make --server required to avoid the confusion.