Update RPM spec file to work with RHEL-8 / python 3 builds
This also simplifies how the python package is built and how
the files are collected in order to create the RPM package.
On RHEL-6 and RHEL-7, the RPMs will end up looking mostly
the same (they will include additional egg-info files).
On RHEL-8 as well as modern Fedoras the package will become
buildable and will work with python 3.